Moved SatelliteType to new signal_defs.h.
