+++ BUILD
@@ -51,6 +51,7 @@ cc_library(
         "src/point_one/fusion_engine/messages/defs.h",
         "src/point_one/fusion_engine/messages/fault_control.h",
         "src/point_one/fusion_engine/messages/measurements.h",
+        "src/point_one/fusion_engine/messages/signal_defs.h",
         "src/point_one/fusion_engine/messages/solution.h",
     ],
     deps = [
